+++ db/migrate/20230912195115_add_required_to_reviews.rb
@@ -1,0 +1,5 @@
+class AddRequiredToReviews < ActiveRecord::Migration[7.0]
+  def change
+    add_column :pull_requests_reviews, :required, :boolean, null: false, default: false
+  end
+end