+++ spec/factories/pull_requests/reviews.rb
@@ -3,7 +3,8 @@
 FactoryBot.define do
   factory :pull_requests_review, class: 'PullRequests::Review' do
     external_id { '1' }
-    review_created_at { DateTime.now }
-    pull_requests_entity
+    review_created_at { nil }
+    pull_request
+    entity
   end
 end