+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- previous insights calculated 1 time per day for faster insight ratio calculations
+
+### Modified
+- skip using PullRequests::Entity model
+- calculating company insights by using repository insights
+
 ## [1.3.8] - 2023-09-11
 ### Added
 - query objects